Implemented the jacobimethod function using the maximum_indices function and rotation function. Running this in the main.cpp file.

+++ project2/include/utils.hpp
@@ -18,7 +18,7 @@ arma::mat rotation(arma::mat A, int N, int k, int l);
 // matrix A. The input arguments k and l are the indices obtain from the
 // maximum_indices function.
 
-arma::mat jacobimethod(arma::mat A, int N, double eps, );
+arma::mat jacobimethod(arma::mat A, int N, int eps);
 // Performs the maximum_indices function and the jacobi rotation in a
 // while loop until max = A(k,l) < 10^eps. Returns the diagonalized
